Sync bitbucket and GitHub

+++ plugins/modules/na_ontap_ipspace.py
@@ -3,18 +3,13 @@
 this is ipspace module
 
 # (c) 2018, NTT Europe Ltd.
-# (c) 2020, NetApp, Inc
+# (c) 2020-2023, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 """
 
 from __future__ import absolute_import, division, print_function
 __metaclass__ = type
 
-ANSIBLE_METADATA = {
-    'metadata_version': '1.1',
-    'status': ['preview'],
-    'supported_by': 'community'
-}
 
 DOCUMENTATION = '''
 ---
@@ -53,7 +48,7 @@ options:
 
 EXAMPLES = """
     - name: Create ipspace
-      na_ontap_ipspace:
+      netapp.ontap.na_ontap_ipspace:
         state: present
         name: ansibleIpspace
         hostname: "{{ netapp_hostname }}"
@@ -61,7 +56,7 @@ EXAMPLES = """
         password: "{{ netapp_password }}"
 
     - name: Delete ipspace
-      na_ontap_ipspace:
+      netapp.ontap.na_ontap_ipspace:
         state: absent
         name: ansibleIpspace
         hostname: "{{ netapp_hostname }}"
@@ -69,7 +64,7 @@ EXAMPLES = """
         password: "{{ netapp_password }}"
 
     - name: Rename ipspace
-      na_ontap_ipspace:
+      netapp.ontap.na_ontap_ipspace:
         state: present
         name: ansibleIpspace_newname
         from_name: ansibleIpspace
@@ -87,9 +82,7 @@ from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils._text import to_native
 import ansible_collections.netapp.ontap.plugins.module_utils.netapp as netapp_utils
 from ansible_collections.netapp.ontap.plugins.module_utils.netapp_module import NetAppModule
-from ansible_collections.netapp.ontap.plugins.module_utils.netapp import OntapRestAPI
-
-HAS_NETAPP_LIB = netapp_utils.has_netapp_lib()
+from ansible_collections.netapp.ontap.plugins.module_utils import rest_generic
 
 
 class NetAppOntapIpspace(object):
@@ -107,20 +100,16 @@ class NetAppOntapIpspace(object):
             argument_spec=self.argument_spec,
             supports_check_mode=True
         )
-
+        self.uuid = None
         self.na_helper = NetAppModule()
         self.parameters = self.na_helper.set_parameters(self.module.params)
 
-        self.rest_api = OntapRestAPI(self.module)
-        if self.rest_api.is_rest():
-            self.use_rest = True
-        else:
-            if HAS_NETAPP_LIB is False:
-                self.module.fail_json(
-                    msg="the python NetApp-Lib module is required")
-            else:
-                self.server = netapp_utils.setup_na_ontap_zapi(module=self.module)
-        return
+        self.rest_api = netapp_utils.OntapRestAPI(self.module)
+        self.use_rest = self.rest_api.is_rest()
+        if not self.use_rest:
+            if not netapp_utils.has_netapp_lib():
+                self.module.fail_json(msg=netapp_utils.netapp_lib_is_required())
+            self.server = netapp_utils.setup_na_ontap_zapi(module=self.module)
 
     def ipspace_get_iter(self, name):
         """
@@ -142,10 +131,9 @@ class NetAppOntapIpspace(object):
             # Error 13073 denotes an ipspace not found
             if to_native(error.code) == "14636" or to_native(error.code) == "13073":
                 return None
-            else:
-                self.module.fail_json(
-                    msg=to_native(error),
-                    exception=traceback.format_exc())
+            self.module.fail_json(
+                msg="Error getting ipspace %s: %s" % (name, to_native(error)),
+                exception=traceback.format_exc())
         return result
 
     def get_ipspace(self, name=None):
@@ -160,20 +148,13 @@ class NetAppOntapIpspace(object):
             name = self.parameters['name']
         if self.use_rest:
             api = 'network/ipspaces'
-            params = None
-            message, error = self.rest_api.get(api, params)
+            query = {'name': name, 'fields': 'uuid'}
+            record, error = rest_generic.get_one_record(self.rest_api, api, query)
             if error:
-                self.module.fail_json(msg=error)
-            if len(message.keys()) == 0:
-                return None
-            elif 'records' in message and len(message['records']) == 0:
-                return None
-            elif 'records' not in message:
-                error = "Unexpected response from %s: %s" % (api, repr(message))
-                self.module.fail_json(msg=error)
-            for record in message['records']:
-                if record['name'] == name:
-                    return record
+                self.module.fail_json(msg="Error getting ipspace %s: %s" % (name, error))
+            if record:
+                self.uuid = record['uuid']
+                return record
             return None
         else:
             ipspace_get = self.ipspace_get_iter(name)
@@ -193,10 +174,10 @@ class NetAppOntapIpspace(object):
         """
         if self.use_rest:
             api = 'network/ipspaces'
-            params = {'name': self.parameters['name']}
-            dummy, error = self.rest_api.post(api, params)
+            body = {'name': self.parameters['name']}
+            dummy, error = rest_generic.post_async(self.rest_api, api, body)
             if error:
-                self.module.fail_json(msg=error)
+                self.module.fail_json(msg="Error provisioning ipspace %s: %s" % (self.parameters['name'], error))
         else:
             ipspace_create = netapp_utils.zapi.NaElement.create_node_with_children(
                 'net-ipspaces-create', **{'ipspace': self.parameters['name']})
@@ -216,13 +197,10 @@ class NetAppOntapIpspace(object):
         :return: None
         """
         if self.use_rest:
-            current = self.get_ipspace()
-            if current is not None:
-                uuid = current['uuid']
-                api = 'network/ipspaces/' + uuid
-                dummy, error = self.rest_api.delete(api)
-                if error:
-                    self.module.fail_json(msg=error)
+            api = 'network/ipspaces'
+            dummy, error = rest_generic.delete_async(self.rest_api, api, self.uuid)
+            if error:
+                self.module.fail_json(msg="Error removing ipspace %s: %s" % (self.parameters['name'], error))
         else:
             ipspace_destroy = netapp_utils.zapi.NaElement.create_node_with_children(
                 'net-ipspaces-destroy',
@@ -243,15 +221,11 @@ class NetAppOntapIpspace(object):
         :return: Nothing
         """
         if self.use_rest:
-            current = self.get_ipspace(self.parameters['from_name'])
-            if current is None:
-                self.module.fail_json(msg="Error renaming ipspace %s" % (self.parameters['from_name']))
-            uuid = current['uuid']
-            api = 'network/ipspaces/' + uuid
-            params = {'name': self.parameters['name']}
-            dummy, error = self.rest_api.patch(api, params)
+            api = 'network/ipspaces'
+            body = {'name': self.parameters['name']}
+            dummy, error = rest_generic.patch_async(self.rest_api, api, self.uuid, body)
             if error:
-                self.module.fail_json(msg=error)
+                self.module.fail_json(msg="Error renaming ipspace %s: %s" % (self.parameters['from_name'], error))
         else:
             ipspace_rename = netapp_utils.zapi.NaElement.create_node_with_children(
                 'net-ipspaces-rename',
@@ -274,8 +248,9 @@ class NetAppOntapIpspace(object):
         """
         current = self.get_ipspace()
         # rename and create are mutually exclusive
-        rename, cd_action = None, None
-        if self.parameters.get('from_name'):
+        rename, modify = None, None
+        cd_action = self.na_helper.get_cd_action(current, self.parameters)
+        if cd_action == 'create' and self.parameters.get('from_name'):
             rename = self.na_helper.is_rename_action(
                 self.get_ipspace(self.parameters['from_name']),
                 current)
@@ -283,19 +258,18 @@ class NetAppOntapIpspace(object):
                 self.module.fail_json(
                     msg="Error renaming: ipspace %s does not exist" %
                     self.parameters['from_name'])
-        else:
-            cd_action = self.na_helper.get_cd_action(current, self.parameters)
-        if self.na_helper.changed:
-            if self.module.check_mode:
-                pass
-            else:
-                if rename:
-                    self.rename_ipspace()
-                elif cd_action == 'create':
-                    self.create_ipspace()
-                elif cd_action == 'delete':
-                    self.delete_ipspace()
-        result = netapp_utils.generate_result(self.na_helper.changed, cd_action)
+            # reset cd_action to None and add name to modify to indicate rename.
+            cd_action = None
+            modify = {'name': self.parameters['name']}
+
+        if self.na_helper.changed and not self.module.check_mode:
+            if rename:
+                self.rename_ipspace()
+            elif cd_action == 'create':
+                self.create_ipspace()
+            elif cd_action == 'delete':
+                self.delete_ipspace()
+        result = netapp_utils.generate_result(self.na_helper.changed, cd_action, modify)
         self.module.exit_json(**result)
